Dumping individual workouts

+++ dump_monthly_listing.py
@@ -5,7 +5,7 @@ from time import sleep
 from dateutil.relativedelta import  relativedelta
 import os
 
-DUMP_DIR='monthly_dumps'
+LISTING_DUMP_DIR='monthly_dumps'
 
 if not 'USER_TOKEN' in os.environ:
     print('USER_TOKEN env not set')
@@ -14,17 +14,16 @@ if not 'USER_ID' in os.environ:
     print('USER_ID env not set')
     quit()
 
-if not os.path.exists(DUMP_DIR):
-    os.mkdir(DUMP_DIR)
+if not os.path.exists(LISTING_DUMP_DIR):
+    os.mkdir(LISTING_DUMP_DIR)
 
 def dump_month(year, month):
-    print(f'Dumping for: {year}-{month}')
     begin_time = datetime(year=year, month=month, day=1)
     end_time = begin_time + relativedelta(months=1)
     # end_time = datetime(year=year, month=month+1, day=1)
     begin_time_str = datetime.strftime(begin_time, '%Y-%m-01T00:00:00.000Z')
     end_time_str = datetime.strftime(end_time, '%Y-%m-01T00:00:00.000Z')
-    print('times', begin_time_str, end_time_str)
+    # print('times', begin_time_str, end_time_str)
     cookies = {
         'USER_TOKEN': os.environ['USER_TOKEN']
     }
@@ -38,11 +37,16 @@ def dump_month(year, month):
 if __name__ == '__main__':
     for year in range(2013, 2020+1):
         for month in range(1, 12+1):
+            print(f'Processing {year}-{month}')
+            json_location = f'{LISTING_DUMP_DIR}/{year}-{month}.json'
+            if os.path.exists(json_location):
+                print(f'{json_location} exists, skipping...')
+                continue
             month_data = {
                 'year': year,
                 'month': month,
                 'data': dump_month(year, month)
             }
-            with open(f'{DUMP_DIR}/{year}-{month}.json', 'w') as fp:
+            with open(json_location, 'w') as fp:
                 json.dump(month_data, fp)
             sleep(1)
